+++ game_patch/multi/server_internal.h
@@ -51,7 +51,7 @@ struct ServerAdditionalConfig
     std::optional<float> max_fov;
     int anticheat_level = 0;
     bool stats_message_enabled = true;
-    bool dynamic_rotation = false;
+    bool random_rotation = false;
     std::string welcome_message;
     bool weapon_items_give_full_ammo = false;
     float kill_reward_health = 0.0f;